add some correction, improve validation, escape script tags

+++ post/views.py
@@ -5,6 +5,8 @@ from django.template.defaultfilters import slugify
 
 from .models import Story, Tag, Comment, Like
 
+import html
+
 
 # Create your views here.
 @login_required
@@ -30,8 +32,8 @@ def index(request):
 @login_required
 def create(request):
     if request.method == 'POST':
-        title = request.POST['title'].strip()
-        content = request.POST['content'].strip()
+        title = request.POST['title']
+        content = validate_text(request.POST['content'])
         tag = request.POST.getlist('tag')
 
         if not title or not content or not tag:
@@ -47,7 +49,7 @@ def create(request):
             slug = slugify(t)
             tag, _ = Tag.objects.get_or_create(name=t, slug=slug)
 
-            tag = tag.strip()
+            print(tag, _)
 
             story.tag.add(tag)
 
@@ -71,10 +73,6 @@ def detail(request, story_id):
         'top_stories': top_stories,
         'most_popular_tag': most_popular_tag
     })
-
-
-def edit(request, story_id):
-    pass
 
 
 @login_required
@@ -113,49 +111,12 @@ def like(request, story_id):
     return HttpResponse(f"<span id='story-like-{like_.story.id}'>{like_.story.get_like_count()}</span>")
 
 
-def validate_title(request):
-    title = request.POST.get('title', '').strip().replace(" ", "")
-
-    if not title:
-        return HttpResponse("<p id='error-title' style='color: red;'>Invalid Title</p>")
-
-    story = Story.objects.filter(title=title).exists()
-
-
-    if not title.isalpha():
-        return HttpResponse("<p id='error-title' style='color: red;'>Invalid Title</p>")
-    if story:
-        return HttpResponse("<p id='error-title' style='color: red;'>Title already exists</p>")
-
-    return HttpResponse("<p id='error-title' style='color: red;'></p>")
-
-
-def validate_content(request):
-    content = request.POST.get('content', '').strip()
-
-    if not content:
-        return HttpResponse("<p id='error-content' style='color: red;'>Invalid content</p>")
-
-
-    if not content.isalpha():
-        return HttpResponse("<p id='error-content' style='color: red;'>Invalid content</p>")
-
-    return HttpResponse("<p id='error-content' style='color: red;'></p>")
-
-
-def validate_tag(request):
-    tag = request.POST.get('tag', '')
-
-    if not tag.isalpha():
-        return HttpResponse("<p id='error-tag' style='color: red;'>Invalid tag</p>")
-
-    return HttpResponse("<p id='error-tag' style='color: red;'></p>")
-
-
 @login_required
 def comment(request, obj_id):
     if request.method == 'POST':
-        content = request.POST.get('comment', '').strip()
+        data = request.POST.get('comment', '').strip()
+
+        content = validate_text(data)
 
         if not content:
             return HttpResponse(status=400)
@@ -166,14 +127,16 @@ def comment(request, obj_id):
             story_id=obj_id
         )
 
-        print(comment_.content)
-
         comment_.save()
 
         return render(request, 'partials/comment.html', {'comment_': comment_})
 
     if request.method == 'DELETE':
         comment_ = Comment.objects.get(id=obj_id)
+
+        if request.user != comment_.author:
+            return HttpResponse(status=403)
+
         comment_.delete()
 
         return HttpResponse("Comment deleted")
@@ -192,3 +155,42 @@ def filter_(request, slug):
     return render(request, 'post/filter.html',
                   {'stories': stories, 'tag': tag, 'tags': tags, 'top_stories': top_stories,
                    'most_popular_tag': most_popular_tag})
+
+
+def validate_title(request):
+    title = request.POST.get('title', '').strip().replace(" ", "")
+
+    if not title:
+        return HttpResponse("<p id='error-title' style='color: red;'>Invalid Title</p>")
+
+    story = Story.objects.filter(title=title).exists()
+
+    if not title.isalnum():
+        return HttpResponse("<p id='error-title' style='color: red;'>Invalid Title</p>")
+    if story:
+        return HttpResponse("<p id='error-title' style='color: red;'>Title already exists</p>")
+
+    return HttpResponse("<p id='error-title' style='color: red;'></p>")
+
+
+def validate_content(request):
+    content = request.POST.get('content', '').replace(" ", "").strip()
+
+    if not content:
+        return HttpResponse("<p id='error-content' style='color: red;'>Invalid content</p>")
+
+    return HttpResponse("<p id='error-content' style='color: red;'></p>")
+
+
+def validate_tag(request):
+    tag = request.POST.get('tag', '')
+
+    if not tag.isalpha():
+        return HttpResponse("<p id='error-tag' style='color: red;'>Invalid tag</p>")
+
+    return HttpResponse("<p id='error-tag' style='color: red;'></p>")
+
+
+def validate_text(content: str) -> str:
+    # Escape the HTML entities to remove any malicious code or script
+    return html.escape(content)
